Script to resolve conflicts and multiple versions.
Often when dealing with the same editors working with multiple devices that
are not always connected to the internet, I end up with multiple versions of
the same file.

This script will attempt to show you each version and let you choose the
"correct" version. It is interactive, so you can just run the script any
time.