Add -l option to by-pass the prompt for a label
With a new name of `note`, this can be used as either a front-end to
a text file of diary entries or just randon notes. The `-l` option lets
you specify the name and by-pass the prompt business.

Now, should I be appending a `_01` to all entries, or just when there
is a conflict?